rename files to help merge
